Tower Buy with Buying and Map Expansion
We can actually buy a tower now!

Adds Expansion Config for map expansion, and with it code to expand the map on a specified wave.

Adds Pause Menu and Tower Detail Panel scripts.

+++ Assets/Scripts/Entities/Tower.cs
@@ -18,7 +18,7 @@ public class Tower : MonoBehaviour
     }
 
     [Header("Statistics"), SerializeField] private TowerType type;
-    [SerializeField] private float price;
+    [SerializeField] private int price;
     [SerializeField] private float damage;
     [SerializeField] private float projectileSpeed;
     [SerializeField] private float rotationSpeed;
@@ -33,7 +33,7 @@ public class Tower : MonoBehaviour
     [Header("Visual"), SerializeField] private Sprite icon;
 
     public TowerType Type => type;
-    public float Price => price;
+    public int Price => price;
     public float Damage => damage;
     public float ProjectileSpeed => projectileSpeed;
     public float RotationSpeed => rotationSpeed;
